5B-Length of Longest Path (Manhattan Tourist)
Find the length of a longest path in a rectangular city.

Given: Integers n and m, followed by an n × (m+1) matrix Down and an
(n+1) × m matrix Right. The two matrices are separated by the "-"
symbol.

Return: The length of a longest path from source (0, 0) to sink (n, m)
in the n × m rectangular grid whose edges are defined by the matrices
Down and Right.